requirements.txt
Flask==3.1.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-Cors==5.0.0
PyJWT==2.10.1
python-dotenv==1.0.1
